Allow empty manifest
During tests on real data we found a rare case when a ubi-repository
can be intentionally empty. But the ubi-manifest didn't differ
between states: "Manifest not found" and "Manifest is empty".

This change allows empty manifest to exist and to be saved into redis
and returned by the application. So now the state "Manifest not found"
clearly means that there is no manifest available for the repository and
therefore empty manifest becomes a valid state.

There is also a little efficiency improvement in if-elif chain in depsolve
task when gathering items for saving to redis.