Improve CI
We add a pre-commit.ci configuration that will automatically add a
commit with code formatting from black and isort.

We update the version of pylint to the latest stable 2.17 version.

We also add a mypy type checker configured so that the checker only
issues warnings and does not fail the whole CI. We configure mypy to be
as lenient as possible.

We also modify the triggers for tox tests, so that the whole suite is
being run only once per PR.