Add poetry for dependency management [RHELDST-19003]
Since poetry is the only tool that reliably solves dependency management
for multiple python versions, we move to installing dependencies via
poetry.

This has the downside that we have to remove rpm-py-installer from
dependencies managed with poetry and install it separately, as poetry
only installs PEP517-compliant packages and rpm-py-installer is not
PEP517 compliant. This should not however cause issues with updating
dependencies as the rpm-py-installer package is virtually unmaintained
at the moment and the last update to the package was in Aug 2022.