Fix keeping n latest rpms for different arches [RHELDST-21855]
Yum repositories can contain rpms with different arches of the same
package. We need to make sure that only the highest and the same version
of rpms are kept.

Previously output contained in some cases pkg.arch_x.v_1 and pkg.arch_y.v2
which is incorrect because of mixed versions and for some workflows
system can end up in the uninstallable state.

This change makes sure that only pkg.arch_x.v_1 and pkg_arch_y.v_1 (if
present) will be in output and it won't end with mixed versions.

This is also fixed for parameter n>=2 which not used currently in practice.