Fix searching for base rpms in debug repos [RHELDST-12605]
The depsolving of rpms is done in two runs, for binary and debug repos
separately. Previously, in both of this runs the search for all modular
filenames was done as well. However, this was useless for the run with
debug repos, as they do not contain modulemd units (in which the modular
rpm filenames are listed).
Also when updating modular `rpm_dependencies` in `modulemd_depsolver`,
the intention (according to the docstring) was to omit source rpms,
since they are searched for separately later in `rpm_depsolver`.
But this was not happening.

This commit intends to fix this and make it more understandable by
couple of changes:
- Pass the list of all modular filenames as an argument to the Depsolver.
The list is then modified in place and therefore is populated from the
binary repos on the first run.
- Make `update_rpm_dependencies` really skip source rpms and remove the
unnecessary division of the rpm dependencies into source and binary
from the depsolver, since we know there won't be any source rpms.
- Put the extension of `in_source_rpm_repos` before the iteration through
cs_repo_map, since it should only be done once.
- Fix some typos, wrong comments.