Fix missing or wrong type annotations [RHELDST-19004]
Make mypy check the incoming code instead of the old
There are a lot of `type: ignore` inline comments because mypy does not
check the type of the results of the Futures, but considers them to be
of type `Future`, resulting in type incompatibility.
Also this commit optimizes the function `_is_blacklisted()` so it is
more clear and easier to type.