Pin the image version for rpm-libs matching python3.11-rpm [RHELDST-22909]
The rpm-libs, which is a builtin package in the base image, is not
matched the python3.11-rpm's installation requirement in the latest
ubi9/python-311 image. And the image building will fail like this:

(app-root) dnf install -y python3.11-rpm
...

Error:
 Problem: conflicting requests
  - nothing provides (rpm-libs(x86-64) >= 4.16.1.3-25 with rpm-libs(x86-64) < 4.16.1.3-26) needed by python3.11-rpm-4.16.1.3-25.1.el9.x86_64 from epel
(try to add '--skip-broken' to skip uninstallable packages or '--nobest' to use not only best candidate packages)

This commit pins the specific version of ubi9/python-311, which is using
the right version of rpm-libs to match python3.11-rpm.