Repository monitoring task [RHELDST-13954]
Added a new task the can do various checks on repositories.
In this change we implemented a check of `last_publish` time. It is
expected that repositories are periodically published and
if the publish hasn't happended in defined time limit, it is logged
so it is possible to intervene, if necessary.

This changes also adds:
* Dockerfile for celery-beat container, new task is set up as
periodic task, so it's required to run celery's scheduler
* updated docker-compose file with addition of celery-beat