Reimplement allowed_repo_groups [RHELDST-22647]
Previously, repo groups for depsolving were hardcoded, often requiring
manual updating. This commit introduces the use of current data from
ubi configs and pulp to determine these groups.

It also changes logging configurations:
  - adds loging.ini configuration file needed to properly log events
  from the app while repo groups are being determined.
  - override celery worker_log_format so that all logs from
  app and workers are in the same format

Finally, it moves all of the gunicorn parameters into the
gunicorn.conf file so that they're all in the same place and the
command to run it is not long.