Stop using repo classes from content_config keys
This commit drops the concept of "repo classes" that defined class names
for the individual ubipop config repos. It could theoretically flatten
the `content_config` dictionary into a list, but keeps it as
a dictionary for backward compatibility.

This is mainly needed for subsequent work of introducing cdn-definitions
as the source of truth for the ubipop config repos instead of
the current `content_config` configuration directive.
The `repo_content_sync` in cdn-definitions only define the git repo
URLs, but doesn't provide any further context, so there is no way
to infer the "repo class" from the git repo alone.

Furthermore, checking whether the repo name matches some arbitrary
predefined string wasn't robust enough, as we plan to introduce
an ubipop config repo that will take care of several distinct products
with content synced to Public CDN which will not fit this assumption.
With various repo names but the same "repo class", this ubipop config
repo couldn't be used by ubi-manifest.